OpeningProof in a separated file to avoid circular dependencies
