+++ verifier_circuit/src/serde/serde_proof.ts
@@ -1,5 +1,6 @@
 import { Group, Proof, Scalar } from "o1js"
-import { OpeningProof, PolyComm } from "../poly_commitment/commitment.js";
+import { PolyComm } from "../poly_commitment/commitment.js";
+import { OpeningProof } from "../poly_commitment/opening_proof.js";
 import { LookupEvaluations, PointEvaluations, ProofEvaluations, ProverCommitments, ProverProof, RecursionChallenge } from "../prover/prover.js"
 import { deserPolyComm, PolyCommJSON, deserGroup, GroupJSON } from "./serde_index.js";
 import { ForeignScalar } from "../foreign_fields/foreign_scalar.js";
